feat: add web database support
