+++ lib/core/providers/providers.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:sqflite/sqflite.dart';
 import '../../data/datasources/local/database_helper.dart';
 import '../../data/repositories/user_repository.dart';
 import '../../data/repositories/user_repository_impl.dart';
