+++ lib/data/datasources/local/conflict_resolver.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:sqflite/sqflite.dart';
+import 'package:sqflite_common/sqlite_api.dart';
 import 'package:logger/logger.dart';
 import '../../../domain/entities/sync_entity.dart';
 import 'database_helper.dart';
