+++ lib/data/datasources/local/sync_manager.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
-import 'package:sqflite/sqflite.dart';
 import 'package:logger/logger.dart';
 import '../../../domain/entities/sync_entity.dart';
 import 'database_helper.dart';
