+++ lib/data/repositories/template_repository.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:sqflite/sqflite.dart';
+import 'package:sqflite_common/sqlite_api.dart';
 import 'package:logger/logger.dart';
 
 import '../../domain/entities/template.dart';
