+++ pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
   # Local Database
   sqflite: ^2.3.0
   sqflite_common_ffi: ^2.3.0+4
+  sqflite_common_ffi_web: ^2.3.0+4
   sqflite_sqlcipher: ^2.1.0+1
   path: ^1.8.3
